api/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
opencv-python
ultralytics
easyocr
Pillow
numpy
sqlalchemy
pytesseract
paddlepaddle
paddleocr
